+++ app/pages/About.tsx
@@ -1,6 +1,6 @@
 import { View, Text, Image, ScrollView, TouchableOpacity } from "react-native";
-import Header from "./components/Header";
-import aboutStyles from "./styles/aboutStyles";
+import Header from "../components/Header";
+import aboutStyles from "../styles/aboutStyles";
 import { Link } from "expo-router";
 
 export default function About() {
@@ -30,11 +30,11 @@ export default function About() {
 
       <View style={aboutStyles.buttonContainer}>
         <TouchableOpacity style={aboutStyles.loginButton} activeOpacity={0.5}>
-          <Link href={"/Login"} style={aboutStyles.loginText}>Login</Link>
+          <Link href={"/auth/Login"} style={aboutStyles.loginText}>Login</Link>
         </TouchableOpacity>
 
         <TouchableOpacity style={aboutStyles.signupButton} activeOpacity={0.5}>
-          <Link href={"/Signup"} style={aboutStyles.signupText}>Sign Up</Link>
+          <Link href={"/auth/Signup"} style={aboutStyles.signupText}>Sign Up</Link>
         </TouchableOpacity>
       </View>
     </View>
