+++ app/pages/Dashboard.tsx
@@ -1,8 +1,8 @@
 import { ScrollView, Text, View } from "react-native";
-import dashBoardStyles from "./styles/dashboardStyles";
-import Footer from "./components/Footer";
-import Header from "./components/Header";
-import UserProfile from "./components/UserProfile";
+import Footer from "../components/Footer";
+import Header from "../components/Header";
+import UserProfile from "../components/UserProfile";
+import dashBoardStyles from "../styles/dashboardStyles";
 
 export default function Dashboard() {
   return (
