+++ app/pages/SplashPage.tsx
@@ -1,7 +1,7 @@
-import React, { useEffect } from "react";
-import { View, Text } from "react-native";
 import { useRouter } from "expo-router";
-import splashStyles from "./styles/splashStyles";
+import React, { useEffect } from "react";
+import { Text, View } from "react-native";
+import splashStyles from "../styles/splashStyles";
 
 export default function SplashPage() {
   const router = useRouter()
